Add ResetReader func to IVFReader
Add ResetReader function for resetting the underlying file
in IVFReader. This is for keeping the stream alive if the file
is being encoded at the same time as it's being streamed.
It takes in a function that returns an io.Reader. A sample
function might look like:
func(bytesRead int64) io.Reader {
	underlyingFile.Seek(bytesRead, io.SeekStart)
	return underlyingFile
}
and you would call it when ivf.ParseNextFrame(...) returns
an error indicating incomplete data.