Increase requirements for Plan-B detection
Before a SessionDescription was assumed to be Plan-B if it contained a
MediaName of `audio` or `video`. This PR Changes that behavior.

If SDPSemantics is UnifiedPlan the SessionDescription must contain a
MediaDescription with two SSRCes to be considered a Plan-B. Before if
the MediaDescription had name of `audio` or `video` it was assumed to be
Plan-B.

If SDPSemantics is PlanB or UnifiedPlanWithFallback it maintains the
existing behavior.

Resolves #2193