Fix potential deadlock in example
At the current state of the code the deadlock is not achievable,
but if 'bufferedAmountLowThreshold' and 'maxBufferedAmount' were to
get changed up, it could happen.
Example: If 'bufferedAmountLowThreshold = 1024',
'maxBufferedAmount = 1280', then if a message of len 999 is sent,
'BufferedAmount()' could return '999', and the condition would still
evaluate to 'true' (999 + 999 > 1280).
Since 'bufferedAmountLowThreshold' was never exceeded, the
'OnBufferedAmountLow' event would never fire, so we're stuck.